chore(infra): batch bump Rust low-risk minors (Phase 4a)
Phase 4a of the dep-update sweep.

Workspace deps:
  axum-tracing-opentelemetry 0.32 → 0.33
  init-tracing-opentelemetry 0.36 → 0.37
  infer 0.16 → 0.19
  smol_str 0.2 → 0.3
  tokio-tungstenite 0.28 → 0.29

Client deps:
  sentry, sentry-tracing 0.47 → 0.48
  sysinfo 0.38 → 0.39 (requires Rust 1.95)

Sub-crate alignment:
  vp8-decoder: thiserror = "1" → thiserror.workspace = true (= "2")

Workspace hygiene:
  tempfile = "3" hoisted to [workspace.dependencies]; server and
  client/src-tauri now use tempfile.workspace = true.

MSRV bump:
  rust-version 1.93 → 1.95 (sysinfo 0.39 requirement). CI uses
  dtolnay/rust-toolchain@stable; stable is now 1.95.0.

Side effects of MSRV bump (rustc 1.95 introduced new clippy lints):
  - websocket_integration.rs ×3: Duration::from_millis(1000) →
    Duration::from_secs(1) (clippy::duration_suboptimal_units)
  - media_processing.rs ×1: inline format arg {error} instead of
    positional (clippy::uninlined_format_args)
  - observability.rs ×2: switch i64+as f64 to native f64 in test
    helper (clippy::needless_type_cast)

Tauri 2.11 schema regeneration:
  client/src-tauri/gen/schemas/{acl-manifests,desktop-schema,
  linux-schema}.json — auto-regenerated by tauri-build from the 2.11
  ACL set; carry-over from Phase 3 (#551) which built only the
  frontend locally.

Per-bump validation (server-side):
  - smol_str From<&str>: server compiles cleanly, no migration needed
  - infer 0.19 MIME db: tests build, runtime validation deferred to CI
  - tokio-tungstenite 0.29 MaybeTlsStream: server compiles cleanly
  - sentry 0.48: client uses sentry-tracing, not the moved tower
    middleware

+++ server/src/admin/observability.rs
@@ -576,9 +576,9 @@ mod tests {
     #[test]
     fn error_rate_calculation() {
         // 5 errors out of 100 requests = 5%
-        let req: i64 = 100;
-        let err: i64 = 5;
-        let rate = err as f64 / req as f64 * 100.0;
+        let req: f64 = 100.0;
+        let err: f64 = 5.0;
+        let rate = err / req * 100.0;
         assert!((rate - 5.0).abs() < f64::EPSILON);
     }
